requirements.txt
python-docx>=0.8.11
pytest>=7.0.0
pytest-cov>=4.0.0
coverage>=7.0.0